pyproject.toml
[project]
name = "producthunt-clone-backend"
version = "0.1.0"
description = "Backend for a ProductHunt clone using FastAPI."
authors = [
    {name = "Hernan Provoste",email = "hernanprovostev@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
